+++ ios/RNSentry.m
@@ -77,6 +77,15 @@ RCT_EXPORT_METHOD(initNativeSdk:(NSDictionary *_Nonnull)options
         reject(@"SentryReactNative", error.localizedDescription, error);
         return;
     }
+    
+    BOOL enableNativeCrashHandling = (BOOL)[mutableOptions valueForKey:@"enableNativeCrashHandling"];
+    
+    if (!enableNativeCrashHandling) {
+        NSMutableArray *integrations = sentryOptions.integrations.mutableCopy;
+        [integrations removeObject:@"SentryCrashIntegration"];
+        sentryOptions.integrations = integrations;
+    }
+    
     [SentrySDK startWithOptionsObject:sentryOptions];
 
     // If the app is active/in foreground, and we have not sent the SentryHybridSdkDidBecomeActive notification, send it.