+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- feat: Support setting maxCacheItems #
+
 ## 3.2.14-beta.2
 
 - feat: Touch events now track components with `sentry-label` prop, falls back to `accessibilityLabel` and then finally `displayName`. #2068