+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 - feat: Support setting maxCacheItems #2102
+- Bump: Sentry Cocoa 7.10.0 #2100
 
 ## 3.2.14-beta.2