+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-- feat: Support setting maxCacheItems #2102
+## Unreleased
+
+- feat: Support enableNativeCrashHandling for iOS #2101
 - Bump: Sentry Cocoa 7.10.0 #2100
+- feat: Support setting maxCacheItems #2102
 
 ## 3.2.14-beta.2